Rewrite without find or findstr (first commit)
Rewrite using only FOR, IF and calling labels. Should be functional up to subtitles, needs testing. Also added "--ui-langue en" to all files and reverted previous master script to work with a variable instead of an array, although that would still need a change in the inner loop (concerning proper invocation of cmd, per the comment to my question on Stack Exchange).